Read the Example_09 data as doubles in the Go and Swift ports
The Go example parsed the population, cell phone and internet numbers
with strconv.ParseFloat(s, 32), and the Swift example with Float(), so
they were rounded to float32 before the division. Java and C# parse them
as doubles and convert only the result to float. Three chart points came
out one float32 step away from Java's, and in Go one circle control point
was written as 511.6 where Java writes 511.59. Both ports now parse
doubles and convert the quotient, as Java does.

+++ Sources/Example_09/main.swift
@@ -119,15 +119,15 @@ public class Example_09 {
             }
 
             var country_name = cols![0].trimmingCharacters(in: .whitespacesAndNewlines)
-            let population = Float(cols![1].filter({ $0 != "," }))
-            let x = Float(cols![5].filter({ $0 != "," }))
-            let y = Float(cols![7].filter({ $0 != "," }).trimmingCharacters(in: .whitespacesAndNewlines))
+            let population = Double(cols![1].filter({ $0 != "," }))
+            let x = Double(cols![5].filter({ $0 != "," }))
+            let y = Double(cols![7].filter({ $0 != "," }).trimmingCharacters(in: .whitespacesAndNewlines))
 
             if population != nil && x != nil && y != nil {
                 let point = Point()
                 point.setText(country_name)
-                point.setX(x! / population!)
-                point.setY((y! / population!) * Float(100.0))
+                point.setX(Float(x! / population!))
+                point.setY(Float(y! / population! * 100.0))
 
                 country_name = country_name.replacingOccurrences(of: " ", with: "_")
                 country_name = country_name.replacingOccurrences(of: "'", with: "_")
